React and snap svg love each other
