+++ .gitignore
@@ -8,3 +8,4 @@ logs/
 tinder_pics/
 node_modules/
 webpack-stats.json
+assets/bundles
