+++ integrations/templates/integrations/twitter_person_detail.html
@@ -13,24 +13,13 @@
     <div class="image-cropper">
         <img src={{ person.avatar }} align="middle"/>
     </div>
+
+    <svg id="svg"></svg>
+
     <h2>{{ person.real_name}}</h2>
-
-<div id="root"></div>
-<div id="container"></div>
-<svg id="svg"></svg>
-{% render_bundle 'main' %}
-    
-    {% if 'TwitterPerson' in person_type %}
-        <h4>@{{ person.username }}</h2>
-        <h6>Sentences: {{ len_sentences }}</h6>
-        <h6>Markov Sentences: {{ len_markov_sentences }}</h6>
-    {% endif %}
-
-    {% if 'LiteraturePerson' in person_type %}
-        <h6>Sentences: {{ len_sentences }}</h6>
-        <h6>Markov Sentences: {{ len_markov_sentences }}</h6>
-    {% endif %}
-
+    <h4>@{{ person.username }}</h2>
+    <h6>Sentences: {{ len_sentences }}</h6>
+    <h6>Markov Sentences: {{ len_markov_sentences }}</h6>
 
     <form action="#" method="get">
         <input name="scrape" type="submit" class="btn btn-success" value="Scrape">
@@ -43,6 +32,8 @@
 </div>
 
 <div class="row small-buffer"></div>
+
+{% render_bundle 'main' %}
 
 {% include "integrations/pool_table.html" with table_title='' pool_form=pool_form %}
 {% include "integrations/sentence_list.html" with table_title='' sentences=sentences %}
