+++ package.json
@@ -4,10 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "node build",
-    "jsx": "jsx -x jsx src .",
-    "watch:jsx": "jsx -w -x jsx src ."
+    "build": "./node_modules/.bin/webpack --config webpack.config.js"
   },
   "author": "",
   "license": "ISC",
@@ -28,9 +25,6 @@
     "webpack-bundle-tracker": "^0.2.0"
   },
   "dependencies": {
-    "art": "^0.10.1",
-    "d3": "^4.9.1",
-    "d3-shape": "^1.2.0",
     "react-svg-morph": "^0.1.10"
   }
 }
